Improved reagent pump model
*The reagent pump model now has conduits connections only visible when actually connected to something on that side.

+++ src/main/java/com/Da_Technomancer/crossroads/API/packets/IIntReceiver.java
@@ -11,6 +11,6 @@ public interface IIntReceiver{
 	 * @param message The actual value being sent. 
 	 * @param sendingPlayer The player who sent the packet. WILL BE NULL IF RECEIVED ON CLIENT SIDE. May also be null if received on server side. 
 	 */
-	public void receiveInt(int identifier, int message, @Nullable EntityPlayerMP sendingPlayer);
+	public void receiveInt(byte identifier, int message, @Nullable EntityPlayerMP sendingPlayer);
 
 }
