+++ src/main/java/com/Da_Technomancer/crossroads/gui/PrototypePortGuiContainer.java
@@ -67,7 +67,7 @@ public class PrototypePortGuiContainer extends GuiContainer{
 		PrototypePortTypes type = typeIndex == 0 ? (in_out.getPressed() == 0 ? PrototypePortTypes.REDSTONE_IN : PrototypePortTypes.REDSTONE_OUT) : typeIndex == 1 ? PrototypePortTypes.ROTARY : typeIndex == 2 ? (in_out.getPressed() == 0 ? PrototypePortTypes.MAGIC_IN : PrototypePortTypes.MAGIC_OUT) : PrototypePortTypes.HEAT;
 		te.setType(type);
 		te.desc = descBar.getText();
-		ModPackets.network.sendToServer(new SendIntToServer(0, side.getPressed() + (type.ordinal() << 3), te.getPos(), te.getWorld().provider.getDimension()));
+		ModPackets.network.sendToServer(new SendIntToServer((byte) 0, side.getPressed() + (type.ordinal() << 3), te.getPos(), te.getWorld().provider.getDimension()));
 		ModPackets.network.sendToServer(new SendStringToServer("desc", descBar.getText(), te.getPos(), te.getWorld().provider.getDimension()));
 		te.getWorld().markBlockRangeForRenderUpdate(te.getPos(), te.getPos());
 	}
